Added support for including in C++ project (don't name-mangle). Added uifcYesNoOpts definition for use with uifc_api.list() and simply Yes/No options.

+++ sbbs3/logout.cpp
@@ -2,13 +2,13 @@
 
 /* Synchronet user logout routines */
 
-/* $Id: logout.cpp,v 1.25 2009/03/20 09:36:20 rswindell Exp $ */
+/* $Id: logout.cpp,v 1.26 2010/03/06 00:13:04 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2009 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2010 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -198,7 +198,7 @@ void sbbs_t::backout()
 		errormsg(WHERE,ERR_OPEN,str,O_RDONLY);
 		return; 
 	}
-	length=filelength(file);
+	length=(long)filelength(file);
 	if((buf=(char *)malloc(length))==NULL) {
 		close(file);
 		errormsg(WHERE,ERR_ALLOC,str,length);
